Download and process future environmental variables made by worldclim version 2.1.

This function downloads future climate data from WorldClim based on CMIP6 climate models and scenarios. The data can be downloaded at various resolutions and clipped to a specific area of interest (AOI) if provided.

Usage

get_worldclim_future(
  var = "bioc",
  res = "30s",
  scenario = "585",
  time_range = "2021-2040",
  gcm = "ACCESS-CM2",
  aoi = NULL,
  retries = 3,
  timeout = 300
)

Arguments

var

Character. Climate variable to download. Available options are:

"bioc" - Bioclimatic variables (19 variables)

"prec" - Precipitation

"tavg" - Average temperature

"tmin" - Minimum temperature

"tmax" - Maximum temperature

Default is "bioc".

res

Character. Spatial resolution of the data. Available options are:

"30s" - ~1 km (30 arc-seconds)

"2.5m" - ~5 km (2.5 arc-minutes)

"5m" - ~10 km (5 arc-minutes)

"10m" - ~20 km (10 arc-minutes)

Default is "30s".

scenario

Character. SSP scenario to use for future climate data. Available options are:

"126" - SSP1-2.6 (low emissions)

"245" - SSP2-4.5 (intermediate emissions)

"370" - SSP3-7.0 (high emissions)

"585" - SSP5-8.5 (very high emissions)

Default is "585".

time_range

Character. Time period for the future scenario. Available options are:

"2021-2040"

"2041-2060"

"2061-2080"

"2081-2100"

Default is "2021-2040".

gcm

Character. The General Circulation Model (GCM) to use. Available options include:

"ACCESS-CM2"

"ACCESS-ESM1-5"

"AWI-CM-1-1-MR"

"BCC-CSM2-MR"

"CanESM5"

"CanESM5-CanOE"

"CMCC-ESM2"

"CNRM-CM6-1"

"CNRM-CM6-1-HR"

"CNRM-ESM2-1"

"EC-Earth3-Veg"

"EC-Earth3-Veg-LR"

"FIO-ESM-2-0"

"GFDL-ESM4"

"GISS-E2-1-G"

"GISS-E2-1-H"

"HadGEM3-GC31-LL"

"INM-CM4-8"

"INM-CM5-0"

"IPSL-CM6A-LR"

"MIROC-ES2L"

"MIROC6"

"MPI-ESM1-2-HR"

"MPI-ESM1-2-LR"

"MRI-ESM2-0"

"UKESM1-0-LL"

Default is "ACCESS-CM2".

aoi

An `sf` or `terra` vector object representing the area of interest (AOI) to which the data will be clipped. Default is `NULL` (no clipping).

retries

Integer. The number of retry attempts for downloading the data in case of failure. Default is 3.

timeout

Numeric. Timeout in seconds for downloading the data. Default is 300 seconds.

Value

A `terra` raster object representing the climate data, optionally clipped to the AOI.

References

Fick, Stephen E., and Robert J. Hijmans. "WorldClim 2: new 1-km spatial resolution climate surfaces for global land areas." International journal of climatology 37.12 (2017): 4302-4315.doi:10.1002/joc.5086

Examples

if (FALSE) { # \dontrun{
library(sf)
library(terra)

nc <- st_read(system.file("shape/nc.shp", package="sf"))
nc <- st_transform(nc, crs = 4326)

climate_future <- get_worldclim_future(var = "tmin",
res = "10m", scenario = "585", time_range = "2021-2040",
gcm = "ACCESS-CM2", aoi = nc)
} # }